+++ software/rt-thread-master/bsp/stm32f40x/User/PropellerControl.h
@@ -20,7 +20,7 @@ void moveLeft(void);  //左移
 void moveRight(void);  //右移
 
 extern PropellerPower_Type  PropellerPower; //推进器推理控制器
-extern int16 Expect_Depth;
+extern int32 Expect_Depth;
 #endif
